Add a manager for authenticators, rework FirebaseAuthenticator.kt and change layout
The manager will be useful to invoke the different signIn, signOut and delete commands and access the connected user. The activity shown on startup is now the login activity. The login layout also has been adapted (this can and will probably change as it is quite simplistic at the moment). FirebaseAuthenticator.kt methods and structure have been reworked (getUser was removed, the database attribute was removed, etc...)