Remove AuthenticatorManager.kt, modify Authenticators and guest options
The AuthenticatorManager.kt made the authentication outside LoginActivity.kt impossible (using the manager) so i decided to remove it.
The authenticators were also adapted because the manager was removed.
Also added a way to skip the LoginActivity.kt when a user is already signed in and added some guest and user options in the AccountFragment.kt.