Resolve problems related to AuthenticatorManager deletion, finalize guest options in AccountFragment.kt
The authenticated user has been moved to DatabaseManager.kt and the code using the user has been changed accordingly.
Resolved the issue when a user logs in/out doesn't refresh the account fragment by making two different fragments for the guest account fragment and the user one.
Add tests for the account fragments and some more tests for LoginActivity.kt